6lowpan: set 6lowpan network and transport header
This is necessary for upcomming patches to calculate the 6lowpan header.
It's a little bit tricky here because a UDP transport header can be
compressed in the network header, so the UDP header is merged in the
6lowpan header. In this case the transport header points to the payload
of the UDP header.

For now it's okay to make it in this way, because we use this to calculate
the 6lowpan header size only.